backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
python-multipart
slowapi
pycryptodome
python-dotenv
cryptography